requirements.txt
streamlit
google-generativeai
langchain
pandas
python-dotenv